html/browsers/windows/browsing-context-first-created.html

<!DOCTYPE html>
<title>HTML Test: Browsing context is first created</title>
<link rel="author" title="Intel" href="http://www.intel.com/" />
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/browsers.html#windows" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

function on_load() {
  var doc = frames[0].document;

  test(function () {
    assert_equals(frames[0].history.length, 1, "The history.length should be 1.");
  }, "The document must be the single document in its browsing context session history");

  test(function () {
    assert_equals(doc.documentURI, "about:blank", "The document's address should be 'about:blank'.");
    assert_equals(doc.URL, "about:blank", "The document's address should be 'about:blank'.");
    assert_equals(doc.contentType, "text/html", "The document should be an HTML document.");
    assert_equals(doc.doctype, null, "The docType of a document without a document type declaration should be null.");
    assert_equals(doc.characterSet, "UTF-8", "The document's encoding should be 'UTF-8'.");
  }, "The document's meta data");

  test(function () {
    assert_equals(doc.readyState, "complete", "The readyState attribute should be 'complete'.");
  }, "The document's status");

  test(function () {
    assert_equals(doc.childNodes.length, 1, "The document must have only one child.");
    assert_equals(doc.documentElement.tagName, "HTML");
    assert_equals(doc.documentElement.childNodes.length, 2, "The HTML element should have 2 children.");
    assert_equals(doc.documentElement.childNodes[0].tagName, "HEAD", "The first child of HTML element should be a HEAD element.");
    assert_false(doc.documentElement.childNodes[0].hasChildNodes(), "The HEAD element should not have child.");
    assert_equals(doc.documentElement.childNodes[1].tagName, "BODY", "The second child of HTML element should be a BODY element.");
    assert_false(doc.documentElement.childNodes[1].hasChildNodes(), "The BODY element should not have child.");
  }, "The document's content");

  test(function () {
    assert_equals(frames[0].parent.document, document);
    assert_equals(doc.referrer, document.URL, "The document's referrer should be its creator Document's address.");
  }, "The document's referrer must be set to the creator document's address");
}

</script>
<iframe onload="on_load()" id="fr"></iframe>
