third_party/WebKit/LayoutTests/imported/wpt/html/browsers/windows/browsing-context-first-created-expected.txt

This is a testharness.js-based test.
PASS Check the history.length of the first created browsing context 
FAIL Check the document's meta data assert_equals: The document's encoding should be 'UTF-8'. expected "UTF-8" but got "windows-1252"
PASS Check the document's status 
PASS Check the document's content 
Harness: the test ran to completion.
